feat(server): add geocoding support to Group creation
This commit adds support for geocoding to the Group creation process. The
geocoding is done using the Mapbox API, and the coordinates are stored in the
Group's geometry field. The Mapbox SDK is added as a dependency.

The createServiceFactory function from the Mapbox SDK is used to create a
geocoding service, which is then used to forward geocode the location provided
in the request body. The resulting coordinates are then added to the Group's
geometry field.

The Group model is updated to include the geometry field, which is a GeoJSON
object containing the type and coordinates of the Group's location.

The seeds file is also updated to use the new geometry field when creating
Groups.